Creates a Controller base class and ArchonController derived class, and moves controller-related functions to that new class. Load ACF functions are working.
