+++ camerad/camera_information.h
@@ -6,7 +6,8 @@ namespace Camera {
   class Information {
     friend class CameraInterface;
     public:
-      Common::FitsKeys userkeys;
+      Common::FitsKeys userkeys;            //!< FITS keys for command line use
+      Common::FitsKeys systemkeys;          //!< FITS keys for system use
 
       std::string base_name;                //!< base image name
 //    std::map<int, std::string> firmware;  //!< firmware file for given controller @TODO support multiple controllers
